Ajoute la quantité en aune (fractions 1/4-1/2-3/4) et Firestore comme base prioritaire
- Stock produit et quantités de vente passent de int à double pour
  supporter le tissu vendu à l'aune, fractionnable par quart.
- Sélecteur dédié (form produit + caisse) pour saisir/vendre en aune
  entières + fraction, avec calcul de prix en direct.
- Affichage du stock et des quantités harmonisé partout (catalogue,
  alertes stock faible, reçus, rapports) via un formatteur commun.
- Ajoute Firestore (via firedart, compatible Windows contrairement au SDK
  officiel) comme base de données prioritaire pour le catalogue produits :
  pull au démarrage, push à chaque écriture, Hive reste le cache
  hors-ligne. Identifiants fournis via --dart-define-from-file, jamais
  commités.

+++ lib/data/models/product_model.dart
@@ -17,7 +17,7 @@ class ProductModel extends HiveObject {
   @HiveField(4)
   String unit;
   @HiveField(5)
-  int stock;
+  double stock;
   @HiveField(6)
   String? color;
   @HiveField(7)
